Implement 'V1Property' which does NOT raise AttributeError
This revisits an old issue relating to attributes which are not set.  The
reason for the original behavior was the documentation on the
SPECIALMETHODS python page about __get__; it says:

    This method should return the (computed) attribute value or raise
    an "AttributeError" exception.

This seemed the correct behavior, and meant that
``hasattr(object, "property")`` returned something useful, which seemed to
confirm that this behavior was "correct" in some way.

However, the documentation of AttributeError[1] has a different wording:

    exception AttributeError
        Raised when an *attribute reference* (see Attribute references) or
        assignment fails. (When an object does not support attribute
        references or attribute assignments at all, TypeError is raised.)

This wording seems subtly different; it seems to refer to the reference,
not the value.  And the wording in the general python data model section on
customizing attribute access[2] seems to imply that being "found in the
class tree for ``self``" would be enough to prevent ``__getattr__`` being
called.  So, by raising ``AttributeError`` the caller cannot suppress them
in the usual way by defining ``__getattr__``.  So, this behavior seems
wrong.

However, there may be callers that depend on the old behavior, so force the
user to explicitly ask for it by requesting a 'V1' property, and a new
property which allows the ``None`` value to be modified.  Properties which
are declared with a falsy ``default`` are half-upgraded to this V1
property; the value is not stored in the instance dictionary; they end up
with 'v1_upgraded' as an instance property.  This is used to maintain the
old effective behavior in JSON round-trip marshaling.

1 - https://docs.python.org/2/library/exceptions.html#exceptions.AttributeError
2 - https://docs.python.org/2/reference/datamodel.html?highlight=attributeerror#object.__getattr__